+++ treecorr/zzcorrelation.py
@@ -98,7 +98,7 @@ class BaseZZCorrelation(Corr2):
     def finalize(self, varz1, varz2):
         """Finalize the calculation of the correlation function.
 
-        The `process_auto` and `process_cross` commands accumulate values in each bin,
+        The `process_auto` and `Corr2.process_cross` commands accumulate values in each bin,
         so they can be called multiple times if appropriate.  Afterwards, this command
         finishes the calculation by dividing each column by the total weight.
 
@@ -279,9 +279,10 @@ class ZZCorrelation(BaseZZCorrelation):
 
     .. note::
 
-        If you separate out the steps of the `process` command and use `process_auto` and/or
-        `process_cross`, then the units will not be applied to ``meanr`` or ``meanlogr`` until
-        the `finalize` function is called.
+        If you separate out the steps of the `Corr2.process` command and use
+        `BaseZZCorrelation.process_auto` and/or `Corr2.process_cross`, then the units will not be
+        applied to ``meanr`` or ``meanlogr`` until the `BaseZZCorrelation.finalize` function is
+        called.
 
     The typical usage pattern is as follows:
 
@@ -314,18 +315,3 @@ class ZZCorrelation(BaseZZCorrelation):
         """Initialize `ZZCorrelation`.  See class doc for details.
         """
         super().__init__(config, logger=logger, **kwargs)
-
-    def finalize(self, varz1, varz2):
-        """Finalize the calculation of the correlation function.
-
-        The `process_auto` and `process_cross` commands accumulate values in each bin,
-        so they can be called multiple times if appropriate.  Afterwards, this command
-        finishes the calculation by dividing each column by the total weight.
-
-        Parameters:
-            varz1 (float):  The variance per component of the first vector field.
-            varz2 (float):  The variance per component of the second vector field.
-        """
-        super().finalize(varz1, varz2)
-
-